Split RTL implementations to files; add preprocessor guards
